Moving enums to a new common file.
