+++ hummingbot/connector/gateway/clob/clob_types.py
@@ -70,18 +70,3 @@ class OrderSide(Enum):
             return HummingBotOrderSide.SELL
         else:
             raise ValueError(f'Unrecognized order side "{self}".')
-
-
-class Chain(Enum):
-    SOLANA = ('solana', 'SOL')
-    ETHEREUM = ('ethereum', 'ETH')
-
-    def __init__(self, chain: str, native_currency: str):
-        self.chain = chain
-        self.native_currency = native_currency
-
-
-class Connector(Enum):
-    def __int__(self, chain: Chain, connector: str):
-        self.chain = chain
-        self.connector = connector
